+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - TASKS : Project planning time line autofill
 - ManufOrder : Hide isConsProOnOperation on manufacturing order when it is false on production process.
 - MANUFACTURING : Hide WORKSHOP menu based on Manfacturing app config.
+- Manuf Order : Forbid to plan a manuf order if the BoM and prodProcess are not applicable.
 
 ## Bug Fixes
 - Configurator : Fix demo data
@@ -93,7 +94,6 @@ in progress manuf order and for bill of material, and add a prorata method for i
 - Translation : Fix english translation of accounting export types.
 - WAITING MODEL : change error message From Waiting model: com.axelor.apps.hr.db.LeaveRequest to Please set the email template to send.
 - ACCOUTING CUTOFF : NPE when stock move not linked to a saleOrder or purchaseOrder
-
 
 ## [5.1.3] - 2019-02-21
 ## Features